improve(designer): GrapesJSBackend 完全実装 + Designer.tsx 統一形に refactor (#815 PR-B)
#815 の段階 2 (PR-B): GrapesJSBackend.renderEditor() を <GjsEditor> +
Canvas + BlocksPanel + RightPanel + DesignSubToolbar をラップする完全実装に
格上げし、Designer.tsx の <GjsEditor> 直接マウントを廃止する。

主要変更:
- GrapesJSBackend.tsx (.ts → .tsx 改名): GrapesJSEditorPane (React コンポーネント)
  を新設し <GjsEditor> ツリー全体を内包。registerRemoteStorage / registerBlocks /
  attachDataItemIdAutoAssign / attachScreenItemsSync / mcpBridge.start / 各種
  listener / canvas-empty 追跡 / theme apply / カスタムブロック CSS 注入 /
  サムネイル生成 を Pane 内で完結させる
- EditorApi 経由で imperative 操作 (applyTheme / reload / refreshCanvas /
  isCanvasEmpty / captureThumbnail / getProjectData / clearUndo) を Designer.tsx
  に expose
- Designer.tsx: <GjsEditor>...</GjsEditor> JSX 廃止、GrapesJSBackend.renderEditor
  を return する形に縮約。state は維持し、handleSave / handleDiscard /
  handleResumeDiscard / handleServerChangeReload / handleLegacyRescueAdopt /
  handleThemeChange を全て EditorApi 経由に書き換え
- onEditor / onReady / canvas-empty 追跡 useEffect / @grapesjs/react import を
  Designer.tsx から削除 (約 540 行縮約)
- handleGrapesChange callback を新設 (Backend からの markDirty signal を受けて
  300ms debounce で API.getProjectData() + updateDraft "screen" を実行)
- handleGrapesReady / handleGrapesServerChanged を新設 (Backend が onReady /
  onServerChanged で発火、Designer.tsx 側で legacy localStorage rescue check と
  ServerChangeBanner 表示制御を行う)
- EditorBackend.test.ts: GrapesJSBackend.renderEditor が React Element を返す
  ことを検証する形に更新 (PR-A での null placeholder テストから格上げ)
- src/test/setup.ts: ResizeObserver mock を追加 (jsdom 環境で @grapesjs/react や
  @measured/puck が require する API を補う)。これにより以前 silent fail していた
  EditorBackend.test.ts / puckScreenValidation.test.ts / serverMtime.test.ts /
  buildConfig.test.ts の 4 suites が通るようになり、Vitest 全体 1486 → 1596 PASS

PR-C (load/save 完全統一) で残作業:
- registerRemoteStorage の autoload 廃止 + autoload: false 化
- editor.load() 全箇所を Backend.load() + editor.loadProjectData() で置換し、
  EditorState.payload を実際の load 経路に乗せる

Refs #815
